+++ src/deployer/ModeDeployer.sol
@@ -2,6 +2,7 @@
 pragma solidity =0.8.25;
 
 import {Deployer} from "./Deployer.sol";
+import {DEPLOYER} from "./DeployerAddress.sol";
 import {MODE_SFS} from "../chains/Mode/IModeSFS.sol";
 
 /// @custom:security-contact security@0x.org
@@ -10,11 +11,11 @@ contract ModeDeployer is Deployer {
 
     constructor(uint256 version) Deployer(version) {
         assert(block.chainid == 34443);
-        if (0x00000000000004533Fe15556B1E086BB1A72cEae.code.length == 0) {
+        if (DEPLOYER.code.length == 0) {
             assert(_implVersion == 1);
             sfsTokenId = MODE_SFS.register(0xf36b9f50E59870A24F42F9Ba43b2aD0A4b8f2F51);
         } else {
-            MODE_SFS.assign(sfsTokenId = MODE_SFS.getTokenId(0x00000000000004533Fe15556B1E086BB1A72cEae));
+            MODE_SFS.assign(sfsTokenId = MODE_SFS.getTokenId(DEPLOYER));
         }
     }