+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Master list of UniV3 forks:
   24. Swapsicle (Algebra-like)
   25. Panko
   26. Swapr (Algebra-like)
+  27. SpookySwap
+  28. Wagmi
 
 ---
 
@@ -54,6 +56,36 @@ Master list of UniV3 forks:
   * Add `DODOV1` action to Mantle
   * Add `DODOV1` action to Polygon
   * Add `DODOV1` action to Scroll
+* Add `rebateClaimer()(address)` function on Mainnet Settlers for gas rebate program
+* Add SolidlyV3 UniV3 fork to Sonic
+* Add Wagmi UniV3 fork to Sonic
+
+## 2025-01-09
+
+### Breaking changes
+
+### Non-breaking changes
+
+* Deploy Settler to Monad testnet chain
+
+## 2024-12-18
+
+### Breaking changes
+
+### Non-breaking changes
+
+* Deploy Settler to Ink chain
+  * Add UniswapV3 UniV3 fork to Ink
+
+## 2024-12-14
+
+### Breaking changes
+
+### Non-breaking changes
+
+* Deploy Settler to Fantom Sonic network
+  * Add UniswapV3 UniV3 fork to Sonic
+  * Add SpookySwap UniV3 fork to Sonic
 
 ## 2024-12-12